Calendar range api works!
It is super duper slow because of how it loads each entry. For example
requesting the range 2016-04-07 with 7 days executes 1 175 queries according to
the webservice logs and only 7 events are included. Requesting 2015-04-07 with
7 days executed 11 819 queries according to webservice logs and 70 events are
included. Executing almost 12k queries for a single week of events isn't going
to scale well when we need to request three months.