Fix login failure response.
Previously, PHP errors were thrown when a user trying to login did not exist in
the database. The errors were returned as HTML snippets in the response. The
specific errors were a missing key (`success` or `succes`) in an array.

Update the webservice auth code to return an array with the `success` parameter
set to false when the username provided cannot be found in the databse.