Use lob.com to verify and then convert the application address to real address.
Add two Freeform hooks that will validate mailing addresses, and report errors
if they are incomplete or incorrect for US addresses.

The validation hook grabs the address fields out of a submission, submits them
to lob.com, and if the result comes back with a "message" it assumes the
address is invalid. The message that lob.com provides is the one presented to
the user.

The save hook runs the address through lob.com and then overwrites whatever the
user inputted with what lob.com returns. In most cases this should be a very
small change like adding the additional 4 numbers to a zip code or rewording
the address.

TODO: provide a way for TeenTix administrators to compare the inputted address
against what lob.com returned.