Fix calendar range organization filter.
Original implementation only filtered the "events" attribute, but "dates"
attribute still had ids for all events from all organzations.

Update webservice calendar extension to filter both events and dates, so
response only includes events for organization set in query string.

there is an opportunity to improve performance in the future by implementing
the filter in the database query. That is a riskier change, so for now we have
a slightly less efficient implementation that pulls all events and then filters
the list down using php code.